fix: comprehensive Tiger Style compliance across 65 files
- Replace .unwrap()/.expect() with proper error handling using snafu context
- Add SAFETY comments to all 14 unsafe blocks
- Replace usize with explicit u32/u64 types for cross-platform consistency
- Add debug_assert! for argument validation and invariants
- Fix naming conventions with unit suffixes (_ms, _bytes) and boolean prefixes
- Add snafu context() to bare ? operators for better error diagnostics
- Decompose compound boolean conditions into early returns
- Add compile-time const assertions for constant relationships
- Refactor long functions using helper prefix convention
- Major improvements to log_subscriber.rs (+401 lines of refactoring)

Total: +740 lines, -422 lines. Build verified passing.

+++ crates/aspen-client-api/src/messages/cluster.rs
@@ -177,7 +177,7 @@ pub struct ClusterTicketResponse {
     /// This node's endpoint ID.
     pub endpoint_id: String,
     /// Number of bootstrap peers in ticket (for combined tickets).
-    pub bootstrap_peers: Option<usize>,
+    pub bootstrap_peers: Option<u32>,
 }
 
 /// Init cluster result response.
